Add selection indicators to image elements also
