+++ demo/app/components/render-slide-elements.ts
@@ -1,4 +1,5 @@
 import { SlideImgElement, SlideTextElement } from '../data/default-slide-data';
+import { queryDataId } from '../utils/dom';
 
 const renderSlideTextElement = (slideElement: SlideTextElement, htmlElement: HTMLElement) => {
   htmlElement.innerHTML = slideElement.text;
@@ -11,8 +12,10 @@ const renderSlideTextElement = (slideElement: SlideTextElement, htmlElement: HTM
   return htmlElement;
 };
 
-const renderSlideImgElement = (slideElement: SlideImgElement, htmlElement: HTMLImageElement) => {
-  htmlElement.src = slideElement.src;
+const renderSlideImgElement = (slideElement: SlideImgElement, htmlElement: HTMLElement) => {
+  const imgElement = queryDataId(htmlElement, 'slide-image-placeholder') as HTMLImageElement;
+  imgElement.src = slideElement.src;
+
   htmlElement.style.position = 'absolute';
   htmlElement.style.left = `${slideElement.position[0]}px`;
   htmlElement.style.top = `${slideElement.position[1]}px`;
