+++ demo/app/index.html
@@ -531,7 +531,10 @@
 
     <template id="slide-image">
       <figure>
-        <img data-id="slide-image-placeholder" />
+        <img
+          data-id="slide-image-placeholder"
+          class="cursor-pointer"
+        />
         <span data-id="slide-figcaption-placeholder"></span>
       </figure>
     </template>
