+++ demo/app/location-tracking/track-slide-elements.ts
@@ -1,87 +1,87 @@
-import Space, { SpaceMember } from '../../../src/Space';
-import { createFragment, queryDataId } from '../utils/dom';
-import { nameToInitials } from '../utils/fake-names';
-import { gradients } from '../utils/gradients';
+import { colors } from '../utils/colors';
+import { HTMLElementManager } from './location-change-handlers';
 
-const slideSelectedClasses = ['mb-[140px]'];
+const slideSelectedClasses = [
+  'outline-2',
+  'outline',
+  `before:content-[attr(data-before)]`,
+  'before:absolute',
+  'before:-top-[22px]',
+  'before:-left-[2px]',
+  'before:px-[10px]',
+  'before:text-sm',
+  'before:text-white',
+  'before:rounded-t-lg',
+  'before:normal-case',
+];
 
-const renderPreviewAvatar = (member: SpaceMember, index: number) => {
-  const fragment = createFragment('#avatar-template');
+const selectSlideElement = (
+  htmlElement: HTMLElement,
+  userName: string,
+  newClientId: string,
+  selfId: string,
+  memberIndex: number,
+) => {
+  htmlElement.setAttribute(`data-client-${newClientId}`, 'true');
 
-  const memberName = member.profileData?.name;
+  const allPresent = htmlElement.getAttribute('data-all-present');
+  htmlElement.setAttribute('data-all-present', `${allPresent ? `${allPresent} ` : ''}${userName}`);
 
-  if (!memberName) {
+  if (!htmlElement.getAttribute('data-before')) {
+    htmlElement.setAttribute('data-before', userName);
+  }
+
+  if (newClientId === selfId) {
+    htmlElement.classList.add(...slideSelectedClasses, 'outline-blue-400', 'before:bg-blue-400');
     return;
   }
 
-  queryDataId(fragment, 'avatar-hover').remove();
+  const color = colors[memberIndex % colors.length];
+  const cssColor = `${color[0]}-${color[1]}`;
+  const outlineColor = `outline-${cssColor}`;
+  htmlElement.classList.add(...slideSelectedClasses, outlineColor, `before:bg-${cssColor}`);
+};
 
-  const initials = queryDataId(fragment, 'name');
-  initials.innerHTML = nameToInitials(member.profileData.name);
+const deselectSlideElement = (
+  htmlElement: HTMLElement,
+  userName: string,
+  oldClientId: string,
+  selfId: string,
+  memberIndex: number,
+) => {
+  htmlElement.removeAttribute(`data-client-${oldClientId}`);
 
-  const innerWrapper = queryDataId(fragment, 'avatar-inner-wrapper');
+  htmlElement.removeAttribute('data-before');
 
-  if (member.isConnected) {
-    innerWrapper.classList.add('bg-gradient-to-b', gradients[index][0], gradients[index][1]);
-    innerWrapper.classList.remove('bg-[##D0D3DC]');
+  const allPresent = htmlElement.getAttribute('data-all-present');
+  const allPresentList = allPresent ? allPresent.split(/\s/) : [];
+  if (allPresentList.length <= 1) {
+    htmlElement.removeAttribute('data-all-present');
   } else {
-    innerWrapper.classList.add('bg-[#D0D3DC]');
-    innerWrapper.classList.remove('bg-gradient-to-b', gradients[index][0], gradients[index][1]);
+    const filteredAllPresent = allPresentList.filter((name) => name !== userName);
+    htmlElement.setAttribute('data-before', filteredAllPresent[0]);
+    htmlElement.setAttribute('data-all-present', filteredAllPresent.join(' '));
   }
 
-  return fragment;
-};
+  const classesToRemove =
+    htmlElement.getAttributeNames().filter((name) => name.startsWith('data-client-')).length > 0
+      ? []
+      : slideSelectedClasses;
 
-const renderPreviewStack = (slideMembers: { member: SpaceMember; i: number }[], slideId: string) => {
-  const ul = document.createElement('ul');
-  ul.classList.add('flex', 'absolute', 'scale-[3]', 'top-[800px]', 'left-[1220px]', 'translate-x-[-100%]');
-  ul.setAttribute('id', `preview-stack-${slideId}`);
-
-  slideMembers.forEach(({ member, i }) => {
-    const avatar = renderPreviewAvatar(member, i);
-    if (avatar) {
-      const li = document.createElement('li');
-      li.classList.add('ml-[-9px]', 'relative');
-      li.appendChild(avatar);
-      ul.appendChild(li);
-    }
-  });
-  return ul;
-};
-
-const updatePreviewAvatarStack = (space: Space, htmlElement: HTMLElement, selfId: string, slideId: string) => {
-  const members = space.getMembers();
-  const slideMembers = members
-    .filter((member) => member.clientId !== selfId)
-    .map((member, i) => ({ member, i }))
-    .filter(({ member }) => member.location && member.location.startsWith(slideId));
-
-  if (slideMembers.length > 0) {
-    htmlElement.classList.add(...slideSelectedClasses);
+  if (oldClientId === selfId) {
+    htmlElement.classList.remove(...classesToRemove, 'outline-blue-400', 'before:bg-blue-400');
+    return;
   }
+  const color = colors[memberIndex % colors.length];
+  const cssColor = `${color[0]}-${color[1]}`;
+  const outlineColor = `outline-${cssColor}`;
 
-  const previewStack = renderPreviewStack(slideMembers, slideId);
-
-  document.getElementById(`preview-stack-${slideId}`)?.remove();
-  htmlElement.appendChild(previewStack);
+  htmlElement.classList.remove(...classesToRemove, outlineColor, `before:bg-${cssColor}`);
 };
 
-const selectSlideElement =
-  (space: Space, slideId: string) =>
-  (htmlElement: HTMLElement, userName: string, newClientId: string, selfId: string) => {
-    updatePreviewAvatarStack(space, htmlElement, selfId, slideId);
-  };
+const slideElementManager: HTMLElementManager = {
+  selector: selectSlideElement,
+  deselector: deselectSlideElement,
+};
 
-const deselectSlideElement =
-  (space: Space, slideId: string) =>
-  (htmlElement: HTMLElement, userName: string, oldClientId: string, selfId: string) => {
-    htmlElement.classList.remove(...slideSelectedClasses);
-    updatePreviewAvatarStack(space, htmlElement, selfId, slideId);
-  };
-
-const createSlideElementManager = (space: Space, slideId: string) => ({
-  selector: selectSlideElement(space, slideId),
-  deselector: deselectSlideElement(space, slideId),
-});
-
-export { renderPreviewStack, createSlideElementManager };
+export { slideElementManager };
